+++ app/src/main/java/com/edibusl/listeatapp/model/repository/UserRepo.java
@@ -4,17 +4,17 @@ package com.edibusl.listeatapp.model.repository;
 import com.edibusl.listeatapp.model.datatypes.User;
 
 public class UserRepo {
-    private int mCurrentUserId;
+    private Long mCurrentUserId;
 
     public UserRepo() {
         //TODO - Load user id
-        mCurrentUserId = 2;
+        mCurrentUserId = 2L;
     }
 
-    public Integer getCurrentUserId() {
+    public Long getCurrentUserId() {
         return mCurrentUserId;
     }
-    public void setCurrentUserId(int userId) {
+    public void setCurrentUserId(Long userId) {
         mCurrentUserId = userId;
     }
 }
